.history/requirements_20250729211552.txt
# Core Flask dependencies
Flask==2.2.2
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.2.1

# Database dependencies
SQLAlchemy==2.0.36
PyMySQL==1.1.1
psycopg2-binary==2.9.9

# Production server
gunicorn==23.0.0

# Security and utilities
Werkzeug==2.2.2
WTForms==3.1.2
Jinja2==3.1.4
MarkupSafe==2.1.5
itsdangerous==2.2.0
blinker==1.8.2
click==8.1.7

# Database migrations
alembic==1.13.3
Mako==1.3.6

# Python utilities
typing_extensions==4.12.2
greenlet==3.1.1
python-dotenv==1.0.0
